Remove deprecated components including AdvancedOptionsPanel, BookingSteps, BookingSummary, ContactForm, DateRangeSelector, DestinationInput, GuestRoomSelector, HotelCard, HotelOffers, HotelResultsFilters, OfferDetailsModal, PopularDestinations, SearchForm, and TravellerForm. Update imports in BookingPage, HotelDetailPage, ResultsPage, and SearchPage to reflect new component structure.
